Restructure App due of Routing Issues
move "fetchData" into App.js;
delete GetData.js;
use props in App.js for routed components;
add more content;
