+++ src/Components/Footer.js
@@ -5,8 +5,7 @@ function Footer (){
 <footer className="bg-light text-center text-lg-start">
   
   <div className="text-center p-3" style= {{backgroundColor: 'fff'}}>
-    © 2022 Copyright:
-    <a className="text-dark" href="https://github.com/Georasam/blog_project.git">github.com/Georasam/blog_project.git</a>
+    <a className="text-dark" href="https://github.com/Georasam/blog_project.git">© 2022 Copyright: github.com/Georasam/blog_project.git</a>
   </div>
   
 </footer>
