+++ src/Components/GameDetails.js
@@ -1,0 +1,9 @@
+function GameDetails () {
+    return (
+        <div>
+            <h1>Game Detail</h1>
+        </div>
+    )
+}
+
+export default GameDetails
